trabalhando na forma de login

+++ itens-js/excluir-material.js
@@ -1,11 +1,13 @@
 
 async function excluirMaterial(){
+
+  if(isLogado == 'true'){
     let codigo = document.getElementById('id-codigo').value;
 
     let filtro = {"filtro":"codigo", "valor":codigo};
     let filtroJSON = JSON.stringify(filtro);
 
-    const req = await fetch(`https://apicontroledematerial.onrender.com/api/item/${filtroJSON}`,{//https://apicontroledematerial.onrender.com
+    const req = await fetch(`https://apicontroledematerial.onrender.com/api/item/${filtroJSON}`,{
          method: "DELETE",
          headers:{"Content-Type": "application/json"},
       });
@@ -13,5 +15,9 @@ async function excluirMaterial(){
     const res = await req.json();
     console.log('resposta: '+res);
 
-    //exibirLista(res.result, 'Devolver outro ítem');
+  }
+  else{
+      if(confirm('Para execultar a operação é necessário estar logado!\nClick em Ok para ser direcionadapara a tela de login.'))
+          window.location.href = '../index.html';
+  }
 }
